freedreno/a6xx: fix hangs with large shaders
We were overflowing instrlen (which is # of groups of 16 instructions)
in a couple dEQP tests, causing gpu hangs:

dEQP-GLES31.functional.ubo.random.all_per_block_buffers.13
dEQP-GLES31.functional.ubo.random.all_per_block_buffers.20

+++ src/gallium/drivers/freedreno/a6xx/fd6_program.c
@@ -239,9 +239,9 @@ struct stage {
 	const struct ir3_shader_variant *v;
 	const struct ir3_info *i;
 	/* const sizes are in units of vec4, aligned to 4*vec4 */
-	uint8_t constlen;
+	uint16_t constlen;
 	/* instr sizes are in units of 16 instructions */
-	uint8_t instrlen;
+	uint16_t instrlen;
 };
 
 enum {
